Delete files from inventory
This commit deletes the entire nixos_inventory_v3.csv file. The reason for this change is to remove outdated inventory data that is no longer needed. By deleting this file, we are cleaning up the repository and ensuring that only relevant files are present. This might be part of a larger refactoring effort to update the inventory system.